Removed "ContactInformation" so now Person owns address, emailaddress, and phone numbers.
